Added tables
Tables drawing is different from the others because as they are 30 elements, their coordinates are not stored in the tables object. Instead, they are represented by playing with for loops and the common properties they all have such as their width, height, their margins in x and y axis, etc.